Implement fight system 2.0: ddd loot exchange after a fight, fight log, retreat opportunity, refactoring and cleanup

+++ app/src/main/java/com/artamonov/millionplanets/gate/GateActivityView.kt
@@ -5,7 +5,7 @@ import com.google.firebase.firestore.DocumentSnapshot
 
 interface GateActivityView {
 
-    fun buyFuel(fuel: Int, money: Int, ship: String?)
+    fun buyFuel(fuel: Long, money: Long, ship: String?)
 
     fun mineDebris()
