+++ app/src/main/java/com/artamonov/millionplanets/model/User.kt
@@ -3,34 +3,34 @@ package com.artamonov.millionplanets.model
 import java.io.Serializable
 
 class User : Serializable {
-    var x: Int = 0
-    var y: Int = 0
+    var x: Long = 0
+    var y: Long = 0
 
-    var sumXY: Int = 0
+    var sumXY: Long = 0
     var ship: String? = null
-    var hp: Int = 0
-    var shield: Int = 0
-    var damage: List<Int>? = null
+    var hp: Long = 0
+    var shield: Long = 0
+    var damage: List<Long>? = null
     var weaponType: List<String>? = null
-    var cargo: Int = 0
-    var scanner_capacity: Int = 0
-    var jump: Int = 0
-    var shipPrice: Int = 0
+    var cargo: Long = 0
+    var scanner_capacity: Long = 0
+    var jump: Long = 0
+    var shipPrice: Long = 0
     var shipClass: String? = null
-    var weaponSlots: Int = 0
+    var weaponSlots: Long = 0
 
-    var money: Int = 0
-    var resource_iron: Int = 0
-    var sectors: Int = 0
+    var money: Long = 0
+    var resource_iron: Long = 0
+    var sectors: Long = 0
     var email: String? = null
     var nickname: String? = null
-    var fuel: Int = 0
+    var fuel: Long = 0
     var type: String? = null
     var moveToObjectName: String? = null
-    var moveToObjectDistance: Int = 0
+    var moveToObjectDistance: Long = 0
     var moveToObjectType: String? = null
 
-    constructor(x: Int, y: Int) {
+    constructor(x: Long, y: Long) {
         this.x = x
         this.y = y
     }
